OBGM-47 Refactor ProductSynonymExcelImporter
